part4.15-4.22

+++ part4/bloglist/controller/blogs.js
@@ -1,6 +1,10 @@
+require('dotenv').config()
 const blogRouter = require('express').Router()
 const { urlencoded } = require('express')
 const Blog = require('../models/blogs')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogRouter.get('/', async (request, response) => {
     const  blogs = await Blog.find({})
@@ -8,10 +12,31 @@ blogRouter.get('/', async (request, response) => {
 })
 
 blogRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
+    const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    console.log("test decoded", decodedToken)
+    
+    if (!decodedToken) {
+      return response.status(400).json({error: 'token invalid'})
+    }
+    
+    const user = await User.findById(decodedToken.id)
+    
+    console.log(user)
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user.id
+    })
+
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    
     response.status(201).json(savedBlog)
-      
 })
 
 blogRouter.put('/:id', async (request, response) => {
@@ -26,6 +51,19 @@ blogRouter.put('/:id', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const body = request.body
+  const decodedToken = jwt.verify(request.token,process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  console.log("blog ",blog)
+  console.log("decoded ",decodedToken)
+  console.log("blog user ",blog.user.toString())
+  console.log("decoded user ",decodedToken.id)
+  if (!decodedToken &&( blog.user.toString() !== decodedToken.user.toString())) {
+    return response.status(400).send('Invalid token')
+  }
+  
+  
+
   const deleted = await Blog.findByIdAndDelete(request.params.id)
   response.status(200).send(deleted)
 })
